refactor: server-style-registry
